create seprate file for each accordion and improve accordion open close funtionality
